feat(cypress): enable component testing and refine spec exclusion patterns
This commit introduces two changes to the Cypress configuration. Firstly, it enables component-based testing by setting the 'component' flag to true in the GitHub Actions workflow file for Cypress. This change allows the tests that specifically target individual UI components to run as part of the CI/CD process.

Secondly, the commit updates the 'excludeSpecPattern' in the Cypress config file by removing the generic pattern '*/*/**/spec.cy.ts'. This alteration ensures that only specified tutorial-related spec files are excluded from the test run, potentially resulting in a broader range of test coverage for component specs.